+++ MasKit/Commands/Open.swift
@@ -9,6 +9,9 @@
 import Commandant
 import Result
 
+private let markerValue = "appstore"
+private let masScheme = "macappstore"
+
 /// Opens app page in MAS app. Uses the iTunes Lookup API:
 /// https://affiliate.itunes.apple.com/resources/documentation/itunes-store-web-service-search-api/#lookup
 public struct OpenCommand: CommandProtocol {
@@ -30,7 +33,19 @@ public struct OpenCommand: CommandProtocol {
     /// Runs the command.
     public func run(_ options: OpenOptions) -> Result<(), MASError> {
         do {
-            guard let result = try storeSearch.lookup(app: options.appId)
+            if options.appId == markerValue {
+                // If no app ID is given, just open the MAS GUI app
+                try systemOpen.run(arguments: masScheme + "://")
+                return .success(())
+            }
+
+            guard let appId = Int(options.appId)
+                else {
+                    print("Invalid app ID")
+                    return .failure(.noSearchResultsFound)
+            }
+
+            guard let result = try storeSearch.lookup(app: appId)
                 else {
                     print("No results found")
                     return .failure(.noSearchResultsFound)
@@ -40,7 +55,7 @@ public struct OpenCommand: CommandProtocol {
                 else {
                     return .failure(.searchFailed)
             }
-            url.scheme = "macappstore"
+            url.scheme = masScheme
 
             do {
                 try systemOpen.run(arguments: url.string!)
@@ -66,7 +81,7 @@ public struct OpenCommand: CommandProtocol {
 }
 
 public struct OpenOptions: OptionsProtocol {
-    let appId: String
+    var appId: String
 
     static func create(_ appId: String) -> OpenOptions {
         return OpenOptions(appId: appId)
@@ -74,6 +89,6 @@ public struct OpenOptions: OptionsProtocol {
 
     public static func evaluate(_ mode: CommandMode) -> Result<OpenOptions, CommandantError<MASError>> {
         return create
-            <*> mode <| Argument(usage: "the app id")
+            <*> mode <| Argument(defaultValue: markerValue, usage: "the app ID")
     }
 }