+++ MasKitTests/Commands/InfoCommandSpec.swift
@@ -41,13 +41,13 @@ class InfoCommandSpec: QuickSpec {
                 storeSearch.reset()
             }
             it("fails to open app with invalid ID") {
-                let result = cmd.run(InfoCommand.Options(appId: "-999"))
+                let result = cmd.run(InfoCommand.Options(appId: -999))
                 expect(result).to(beFailure { error in
                     expect(error) == .searchFailed
                 })
             }
             it("can't find app with unknown ID") {
-                let result = cmd.run(InfoCommand.Options(appId: "999"))
+                let result = cmd.run(InfoCommand.Options(appId: 999))
                 expect(result).to(beFailure { error in
                     expect(error) == .noSearchResultsFound
                 })
@@ -57,7 +57,7 @@ class InfoCommandSpec: QuickSpec {
                 let output = OutputListener()
                 output.openConsolePipe()
 
-                let result = cmd.run(InfoCommand.Options(appId: result.trackId.description))
+                let result = cmd.run(InfoCommand.Options(appId: result.trackId))
 
                 expect(result).to(beSuccess())
                 // output is async so need to wait for contents to be updated